refactor: 폴더 구조 정상화 (2)
