+++ packages/service/src/page/main/content/loading.tsx
@@ -1,8 +1,8 @@
 import { getScopedI18n } from '@/locales/server';
-import { MainLastSection } from '@/widget/main/section/neighborhood';
 
 import { MainCategorySectionFallback } from '../section/category/fallback';
 import { MainCardSectionFallback } from '../section/fallback';
+import { MainLastSection } from '../section/neighborhood';
 import { MainUpcommingScheduleFallback } from '../section/upcoming/fallback';
 
 export async function MainContentLoadingFallback() {
