+++ packages/service/src/page/main/main.css.ts
@@ -1,4 +1,3 @@
-import { rem } from '@moeasy/storybook/utils/css';
 import { headerHeight } from '@moeasy/storybook/utils/styles/global.css';
 
 import { style } from '@vanilla-extract/css';
@@ -8,9 +7,4 @@ export const main = style({
   flexDirection: 'column',
   alignItems: 'center',
   minHeight: `calc(100vh - ${headerHeight})`,
-});
-
-export const commonCard = style({
-  minWidth: rem(327),
-  height: rem(171),
 });
