+++ packages/service/src/widget/card/fallback.tsx
@@ -2,12 +2,11 @@ import { sprinkles } from '@/shared/style/sprinkles/index.css';
 
 import { CardHeader, CardThumbnail, CardWrapper } from '@moeasy/storybook/ui/card/compound-card';
 
-import { commonCard } from '../../main.css';
 import * as styles from './card.css';
 
 export function MainCommonCardFallback() {
   return (
-    <CardWrapper className={commonCard}>
+    <CardWrapper className={styles.commonCard}>
       <CardThumbnail />
       <CardHeader />
       <div className={styles.cardInfo}>
