+++ packages/service/src/widget/card/schedule/schedule.css.ts
